Make the classes for object-orientated in extra files.
