+++ helper_functions/game_utils.py
@@ -1,0 +1,5 @@
+from flask import session
+
+
+def get_bet_amount():
+    return session.get('bet_amount', 0)
